+++ packages/ridb-core/src/plugin/defaults/mod.rs
@@ -20,12 +20,12 @@ impl DefaultsPlugin {
         let plugin_clone1 = plugin.clone();
         let create_hook = Closure::wrap(Box::new(move |schema, _migration, document| {
             // Add logging for debugging
-            Logger::debug(&"Creating document with defaults".into());
+            Logger::debug("DefaultsPlugin", &"Creating document with defaults".into());
             let result = plugin_clone1.clone().add_defaults(schema, document);
             if result.is_ok() {
-                Logger::debug(&"Document created successfully".into());
+                Logger::debug("DefaultsPlugin", &"Document created successfully".into());
             } else {
-                Logger::debug(&"Failed to create document".into());
+                Logger::debug("DefaultsPlugin", &"Failed to create document".into());
             }
             result
         }) as Box<dyn Fn(JsValue, JsValue, JsValue) -> Result<JsValue, JsValue>>);
@@ -36,7 +36,7 @@ impl DefaultsPlugin {
     
 
     pub(crate) fn add_defaults(&self, schema: JsValue, document: JsValue) -> Result<JsValue, JsValue> {
-        Logger::debug(&"Adding defaults to document".into());
+        Logger::debug("DefaultsPlugin", &"Adding defaults to document".into());
         let schema = Schema::create(schema)?;
 
         let properties = schema.properties.clone();
@@ -45,7 +45,7 @@ impl DefaultsPlugin {
             if current_value.is_null() || current_value.is_undefined() {
                 let has_default = prop.default.is_some();
                 if has_default {
-                    Logger::debug(&format!("Setting default for key: {}", key).into());
+                    Logger::debug("DefaultsPlugin",&format!("Setting default for key: {}", key).into());
                     Reflect::set(
                         &document, 
                         &JsValue::from_str(&key), 
@@ -54,7 +54,7 @@ impl DefaultsPlugin {
                 }
             }
         }
-        Logger::debug(&"Defaults added successfully".into());
+        Logger::debug("DefaultsPlugin",&"Defaults added successfully".into());
         Ok(document)
     }